requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
blessings==1.7
bpython==0.20.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
curtsies==0.3.4
dadata==20.7.0
Django==3.1.4
django-cors-headers==3.6.0
django-dirtyfields==1.4.1
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-sberbank==0.2.31
django-silk==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gprof2dot==2019.11.30
greenlet==0.4.17
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
idna==2.10
isort==5.6.4
Jinja2==2.11.2
jsonfield2==4.0.0.post0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary
pycodestyle==2.6.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
